Refactor settings to a separate file
[MAILPOET-2452]
